+++ internal/storage/messages.go
@@ -112,23 +112,29 @@ func Store(body *[]byte) (string, error) {
 		return "", err
 	}
 
-	// extract tags from body matches
-	rawTags := findTagsInRawMessage(body)
-	// extract plus addresses tags from enmime.Envelope
-	plusTags := obj.tagsFromPlusAddresses()
-	// extract tags from X-Tags header
-	xTags := tools.SetTagCasing(strings.Split(strings.TrimSpace(env.Root.Header.Get("X-Tags")), ","))
-	// extract tags from search matches
-	searchTags := tagFilterMatches(id)
+	// extract tags using pre-set tag filters, empty slice if not set
+	tags := findTagsInRawMessage(body)
 
-	// combine all tags into one slice
-	tags := append(rawTags, plusTags...)
-	tags = append(tags, xTags...)
-	// sort and extract only unique tags
-	tags = sortedUniqueTags(append(tags, searchTags...))
+	if !config.TagsDisableXTags {
+		xTagsHdr := env.Root.Header.Get("X-Tags")
+		if xTagsHdr != "" {
+			// extract tags from X-Tags header
+			tags = append(tags, tools.SetTagCasing(strings.Split(strings.TrimSpace(xTagsHdr), ","))...)
+		}
+	}
 
+	if !config.TagsDisablePlus {
+		// get tags from plus-addresses
+		tags = append(tags, obj.tagsFromPlusAddresses()...)
+	}
+
+	// extract tags from search matches, and sort and extract unique tags
+	tags = sortedUniqueTags(append(tags, tagFilterMatches(id)...))
+
+	setTags := []string{}
 	if len(tags) > 0 {
-		if err := SetMessageTags(id, tags); err != nil {
+		setTags, err = SetMessageTags(id, tags)
+		if err != nil {
 			return "", err
 		}
 	}
@@ -144,7 +150,7 @@ func Store(body *[]byte) (string, error) {
 	c.Attachments = attachments
 	c.Subject = subject
 	c.Size = size
-	c.Tags = tags
+	c.Tags = setTags
 	c.Snippet = snippet
 
 	websockets.Broadcast("new", c)
@@ -593,7 +599,9 @@ func DeleteMessages(ids []string) error {
 		}
 	}
 
-	err = tx.Commit()
+	if err := tx.Commit(); err != nil {
+		return err
+	}
 
 	dbLastAction = time.Now()
 	addDeletedSize(int64(totalSize))