+++ server/pop3/pop3_test.go
@@ -67,7 +67,7 @@ func TestPOP3(t *testing.T) {
 		return
 	}
 
-	count, size, err = c.Stat()
+	count, _, err = c.Stat()
 	if err != nil {
 		t.Errorf(err.Error())
 		return
@@ -349,7 +349,7 @@ func insertEmailData(t *testing.T) {
 			t.Fail()
 		}
 
-		if err := storage.SetMessageTags(id, []string{fmt.Sprintf("Test tag %03d", i)}); err != nil {
+		if _, err := storage.SetMessageTags(id, []string{fmt.Sprintf("Test tag %03d", i)}); err != nil {
 			t.Log("error ", err)
 			t.Fail()
 		}